Autocomplete: group `getInlineCompletions` unit tests
