+++ vscode/src/completions/get-current-doc-context.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
 
 import { getCurrentDocContext } from './get-current-doc-context'
-import { documentAndPosition } from './testHelpers'
+import { documentAndPosition } from './test-helpers'
 
 describe('getCurrentDocContext', () => {
     it('returns `docContext` for a function block', () => {
