+++ vscode/src/completions/request-manager.test.ts
@@ -4,7 +4,7 @@ import { vsCodeMocks } from '../testutils/mocks'
 
 import { Provider } from './providers/provider'
 import { RequestManager, RequestManagerResult, RequestParams } from './request-manager'
-import { documentAndPosition } from './testHelpers'
+import { documentAndPosition } from './test-helpers'
 import { getNextNonEmptyLine, getPrevNonEmptyLine } from './text-processing'
 import { Completion } from './types'
