+++ vscode/src/completions/vscodeInlineCompletionItemProvider.test.ts
@@ -11,7 +11,7 @@ import { vsCodeMocks } from '../testutils/mocks'
 import { getInlineCompletions, InlineCompletionsResultSource } from './getInlineCompletions'
 import * as CompletionLogger from './logger'
 import { createProviderConfig } from './providers/anthropic'
-import { documentAndPosition } from './testHelpers'
+import { documentAndPosition } from './test-helpers'
 import { InlineCompletionItem } from './types'
 import { InlineCompletionItemProvider } from './vscodeInlineCompletionItemProvider'
